py3_azure/func_app/requirements.txt
azure-storage-blob == 2.1.0
azure-functions
